pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "capydocs"
version = "0.2.0"
description = "A lightweight web-based markdown document manager for your repository"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "mjk" }]
keywords = ["markdown", "docs", "editor", "web"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.23.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=2.0.0",
]

[project.optional-dependencies]
ai = [
    "openai>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
    "ruff>=0.4.0",
    "build>=1.0.0",
]

[project.scripts]
capydocs = "capydocs.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
capydocs = ["static/**/*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["capydocs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
